exercise 4.22

+++ controllers/blogs.js
@@ -23,6 +23,10 @@ blogsRouter.get('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken || !decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
   const result = await Blog
     .findByIdAndDelete(request.params.id)
   if (result) {
@@ -60,10 +64,7 @@ blogsRouter.post('/', async (request, response, next) => {
     return response.status(401).json({ error: 'token invalid' })
   }
 
-  const user = await User.findById(decodedToken.id)
-  if (!user) {
-    return response.status(400).json({ error: 'UserId missing or not valid' })
-  }
+  const user = request.user
 
 
   const blog = new Blog({
